pyproject.toml
[project]
name = "openregulations"
version = "0.1.0"
description = "AI-powered analysis of federal regulatory comments"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}

dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "supabase>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
modal = [
    "modal>=0.56.0",
]

[project.scripts]
sync = "src.jobs.sync:main"
analyze = "src.jobs.analyze:main"
embed = "src.jobs.embed:main"
report = "src.jobs.report:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
